Fix user_roles role_id FK type to match roles.id

+++ backend/database/migrations/2026_03_14_000100_create_user_roles_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
         Schema::create('user_roles', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('role_id');
+            $table->unsignedInteger('role_id');
             $table->enum('status', [
                 UserRole::STATUS_ACTIVE,
                 UserRole::STATUS_ONBOARDING,
